Move cleanup to OnCoreMapEnd Moce CGameEventListener to its own file (same case)
